requirements.txt
fastapi
uvicorn
celery
redis
pytesseract
networkx
sqlalchemy
faker
python-multipart
opencv-python
pydantic
torch
torchvision
torchaudio
pillow
requests
pandas
passlib[bcrypt]
python-jose[cryptography]
bcrypt==3.2.2
